+++ packages/api/lib/writeRecords/write-granules.js
@@ -712,7 +712,7 @@ const writeGranuleRecordAndPublishSns = async ({
  * @param {boolean} [granule.published] - published to cmr
  * @param {string} [granule.pdrName] - pdr name
  * @param {string} [granule.provider] - provider
- * @param {boolean} [granule.active] - Flag for whether the granule is active or hidden
+ * @param {boolean} [granule.activeStatus] - Flag for whether the granule is active or hidden
  * @param {Object} [granule.error = {}] - workflow errors
  * @param {string} [granule.createdAt = new Date().valueOf()] - time value
  * @param {string} [granule.timestamp] - timestamp
@@ -739,7 +739,7 @@ const writeGranuleRecordAndPublishSns = async ({
 const writeGranuleFromApi = async (
   {
     granuleId,
-    active,
+    activeStatus,
     collectionId,
     createdAt,
     status,
@@ -803,7 +803,7 @@ const writeGranuleFromApi = async (
       granuleId,
       cmrLink,
       producerGranuleId,
-      active,
+      activeStatus,
       published: publishedValue,
       createdAt: defaultCreatedAt,
       error: defaultSetError,
@@ -830,7 +830,7 @@ const writeGranuleFromApi = async (
       cmrTemporalInfo,
       cmrUtils,
       collectionId,
-      active,
+      activeStatus,
       createdAt: defaultCreatedAt,
       duration,
       executionUrl: execution,
@@ -1056,7 +1056,7 @@ const writeGranulesFromMessage = async ({
           producerGranuleId: granule.producerGranuleId || granule.granuleId,
         },
         executionUrl,
-        active: granule.active,
+        active: granule.activeStatus,
         collectionId,
         provider: provider.id,
         files,