+++ packages/types/api/granules.d.ts
@@ -35,7 +35,7 @@ export type ApiGranuleRecord = {
   producerGranuleId: string
   status: GranuleStatus
   updatedAt: number
-  active: string
+  activeStatus?: string
   cmrLink?: string
   duration?: number
   error?: Object
@@ -57,7 +57,7 @@ export type ApiGranule = {
   collectionId: string
   archived: boolean
   producerGranuleId: string
-  active: string
+  activeStatus?: string
   cmrLink?: string | null
   createdAt?: number | null
   duration?: number | null